Writing matrix to files
