+++ tools/matrix-writer.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#ifndef MATRIX_WRITER_HPP
+#define MATRIX_WRITER_HPP
+
+#include <stdlib.h>
+
+int write_matrix_to_file(double **A, size_t n, char *filename);
+
+#endif
